Used yt-search library to search for song videos.
Using the first in the result to play the video.
I am also embedding the video that is being played.
Implemented the resume functionality that was missing... If you paused, then you could not continue.

The code needs to be refactored, and there are many ways the program will crash. E.g passing an empty play command...

Also need to implement a queue, and find ways of initializing some global variables at the start of the program
instead of  using undefined variables.